pyproject.toml
[project]
name = "streamlit-mialbarinov"
version = "0.1.0"
description = ""
authors = [
    {name = "Mikhail",email = "mialbarinov@edu.hse.ru"}
]
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "streamlit (>=1.52.2,<2.0.0)",
    "requests (>=2.32.5,<3.0.0)",
    "httpx (>=0.28.1,<0.29.0)",
    "python-dotenv (>=1.2.1,<2.0.0)"
]


[tool.ruff]
target-version = "py312"
line-length = 120
exclude = [
    ".git",
    ".mypy_cache",
    ".ruff_cache",
    ".venv",
    "__pycache__",
    "data/analysis",
]

[tool.ruff.lint]
select = ["E", "F", "I", "W", "B"]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
